Download Census 2020 data files from FTP - For Island Areas VI GU MP AS

Download Census 2020 data files from FTP - For Island Areas VI GU MP AS

Usage

census2020_download_islandareas(
  folder = NULL,
  mystates = c("VI", "GU", "MP", "AS"),
  allstates = c("VI", "GU", "MP", "AS"),
  baseurl = "https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/",
  urlmiddle = "demographic-and-housing-characteristics-file/",
  zipnames_suffix = "2020.dhc.zip"
)

Arguments

folder

Default is a tempdir. Folder and subfolder for data are created if they do not exist.

mystates

Character vector of 2 letter abbreviations, optional,

Default is VI GU MP AS

allstates

Default is VI GU MP AS

baseurl

default is the FTP folder with the data

Value

Effect is to download and save locally a number of data files.

See also

Examples

if (FALSE) {
 # 
 }